--- Schedule.Web/Helpers/TokenParameters.cs
@@ -1,36 +1,0 @@
-using Microsoft.IdentityModel.Tokens;
-using System;
-using System.Text;
-
-namespace Schedule.Web.Helpers
-{
-    public class TokenParameters
-    {
-        public static TokenValidationParameters GetTokenValidationParameters(string secretKey)
-        {
-            var signingKey = new SymmetricSecurityKey(Encoding.ASCII.GetBytes(secretKey));
-            return new TokenValidationParameters
-            {
-                // The signing key must match!
-                ValidateIssuerSigningKey = true,
-                IssuerSigningKey = signingKey,
-
-                // Validate the JWT Issuer (iss) claim
-                ValidateIssuer = true,
-                ValidIssuer = "ExampleIssuer",
-
-                // Validate the JWT Audience (aud) claim
-                ValidateAudience = true,
-                ValidAudience = "ExampleAudience",
-
-                // Validate the token expiry
-                ValidateLifetime = true,
-
-                // If you want to allow a certain amount of clock drift, set that here:
-                ClockSkew = TimeSpan.Zero,
-
-                SaveSigninToken = true
-            };        
-        }
-    }
-}
